refactor(product-catalogs): replace Blueprint Excel services with YAML processor
Remove Excel-based blueprint generation (exporter, template, processor)
and add BlueprintYamlProcessorService that parses the product_profiles.yaml
format directly, using the same PG COPY upsert strategy.

Joel Cano <jcano@merkur.la>
